refactor(desktop): add web search provider adapters
